pyproject.toml
[project]
name = "appointment-booking-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=1.2.10",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.1",
    "langchain-ollama>=1.0.1",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.48",
    "streamlit>=1.55.0",
]
